Renamed Equals to Comparison and moved Binary and Comparison Ops to their own files
